+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Feat #299: Add 'description' - 'prefix' - 'displayed as' - 'multiple' - 'relationship_id' columns to the external_link_type table
+- Feat #299: Admin pages to manage externalLink types
+
 ## v4.4.16 - 2025-08-15 - 6e04c0b9e -
 
 - Fix #2428: AdminUserController should target AdminUserController when validating or rejecting a claim